Physics' law attempt
I need to work with vectors to store the positions. Adittionally, I need to calculate the velocity based on the past, on the present and on the future. I can do so by making a backward, forward or midpoint calculation. But first, how can I update the positions (in a loop) and calculate these velocities (in another loop?).